+++ src/routes/formatt/constants.ts
@@ -1,1 +1,0 @@
-export { DEFAULT_FORMAT_PARAMS as defaultFormatParams } from '$lib/config/formatting';